+++ CHANGELOG.md
@@ -1,12 +1,14 @@
-# CHANGELOG of Itamae::Kitchen
+CHANGELOG of Itamae::Kitchen
 
 ## 0.0.2
 
+- cleanup code
 - cleanup sandbox
 - cleanup remote root_path before converge
 - option with_ohai
 - option sudo_command
 - option itamae_option
+- plugin install by embedded gem (every run)
 
 ## 0.0.1